+++ CHANGELOG.md
@@ -10,3 +10,11 @@ Changelog
 ### 1.2.0
 - Added support for automatic signal inheritance from base classes
 
+### 1.2.1
+- Fixed automatic signal inheritance for multiple base classes
+
+### 1.2.2
+- Changed connect to raise error instead of warning upon undefined signal
+
+### 1.2.3
+- Added explicit metaclass for Connectable needed for resolving metaclass conflicts